SMART... structers ! ! !! ! Progressive generation !!
FUCKING FINALLY !!!
Instead of freezing like a dumbass when generating the structures of whole MegaChunks at the same time, it now generates individual structures over time in MegaChunks !
--> The 8 neighbours of loaded MegaChunks are kept loaded as ExtraMegaChunks
-> When ExtraMegaChunks are not fully generated and are close enough to player pos (4 manhattan in megachunk pos), they are susceptible to have their structures generated. The screen focuses on a SINGLE one at a time : it will pick an extraloadedMegaChunk, keep it in memory, and generate one structure of it each each frame until it's done  generating all of its structures, after which it's marked as fully generated.
-> When a MegaChunk is getting its structures generated, it's pos is stored in generatingMegaChunk, and the extraLoadedChunks that get extraLoaded because of structure generation CANNOT GET UNLOADED. So that if it generates multiple structures at the same tile they aren't loaded unloaded loaded... to be more efficient.

-----> How does it know which structures to generated over time ? MegaChunks now have a to do list of all structures to spawn : when a MegaChunk not doneGeneratingStructures is targeted for structures generation, it will put in a list all the structures to generate (position, seed, type...). If game is closed/megaChunk gets unloaded, the list is saved in the megaChunkJson, so nothing is lost (idk if it that saving works lol but it should lol).

--> When an extraLoadedMegaChunk is getting promoted to being fully loaded, if it doesn't have all its structures generated, they will all be forcefully generated (to not fuck up shit), might freeze, but will rarely happen now ! ! So basically if emergency it will just as like before ig.

----> Debug loaded chunks/megachunks minimap updated !!
-> Extraloaded megaChunks now appear, in dark red and blinking, smoothly over 1s, but only in the dimension the player is in ! ! !
-> The generatingMegaChunk (the one currently generating structures in the background) is in brown instead, still blinking if in player dimension only.
-> The extraLoadedChunks are not in light green instead of purple, looks better and they're distinguishable from structureloadedchunks now ! ! !!

--------------------------------------->>>>>>>> IT WORKS !!!!!!!!!!! But when generating structures in the background it lags like fucking crazy... like slows down by 2... still better than the 3 seconds freeze, and due to chunk generation mostly ! Will have to improve that ughhh

Misc : LCGs now have an Abs each time in case... so you never get negative seeds anymore, nice nice ougermldk

+++ Plants.cs
@@ -127,7 +127,7 @@ namespace Cave
             }
             public int getplantRandValue(int mod = -1)
             {
-                randValue = (int)cashInt((seed, 7, 13), 0);
+                randValue = cashInt((seed, 7, 13), 0);
                 if (mod == 0) { return 0; }
                 return mod > 0 ? randValue % mod : randValue;
             }
@@ -463,7 +463,7 @@ namespace Cave
             }
             public int getRandValue(int valueModifier, int modulo = -1)
             {
-                int randy = (int)cashInt((valueModifier, 7, 13), seed);
+                int randy = cashInt((valueModifier, 7, 13), seed);
                 if (modulo == 0) { return 0; }
                 return modulo > 0 ? randy % modulo : randy;
             }
